Fix broken HTTP support.
HTTP image support was completely broken. Added check for updated
Http.IncomingMessage.headers property in addition to checking for
getHeader(). Also explicitly set the encoding option to 'null' so that
the body value of the respons is always returned as a buffer preventing
the parsing libraries from failing to read Encoding headers properly.

Added three test cases to verify the fixes (one for each img type). They
are pointed to the github raw test images, which is safe for now.